fix: remove illegal sequence termination and increased holding malus

+++ main.py
@@ -17,15 +17,18 @@ flags.DEFINE_boolean('clean_run', False, 'Clean the checkpoints and results fold
 def main(_):
     if FLAGS.clean_run:
         for folder in ['checkpoints', 'results']:
-            for filename in os.listdir(folder):
-                file_path = os.path.join(folder, filename)
-                try:
-                    if os.path.isfile(file_path) or os.path.islink(file_path):
-                        os.unlink(file_path)
-                    elif os.path.isdir(file_path):
-                        shutil.rmtree(file_path)
-                except Exception as e:
-                    print('Failed to delete %s. Reason: %s' % (file_path, e))
+            if os.path.exists(folder):
+                for filename in os.listdir(folder):
+                    file_path = os.path.join(folder, filename)
+                    try:
+                        if os.path.isfile(file_path) or os.path.islink(file_path):
+                            os.unlink(file_path)
+                        elif os.path.isdir(file_path):
+                            shutil.rmtree(file_path)
+                    except Exception as e:
+                        print('Failed to delete %s. Reason: %s' % (file_path, e))
+            else:
+                print('Nothing to clean in %s' % folder)
 
     console = Console()
